Update - In-Depth Study of the filter Method
This commit brings an in-depth study of the filter method in arrays, exploring concepts and specific applications in the Advanced Arrays section. Annotations have been adjusted to highlight the effective use of the filter method in different advanced scenarios. This is another step in the exploration of advanced array methods.