Update - In-Depth Study of the map Method
This commit presents an in-depth study of the map method in arrays, exploring concepts and specific applications in the Advanced Arrays section. Annotations have been adjusted to highlight the effective use of the map method in different advanced scenarios. This is another step in the exploration of advanced array methods.