Update - In-Depth Study of the reduce Method
This commit brings an in-depth study of the reduce method in arrays, exploring concepts and specific applications in the Advanced Arrays section. Annotations have been adjusted to highlight the effective use of the reduce method in different advanced scenarios of array reduction. This is another step in the exploration of advanced array methods.