Update - Study on the Combination of Filter, Map, and Reduce in Arrays
This commit presents a study on the combination of filter, map, and reduce in arrays, exploring how these methods can be used together to perform advanced data transformation and reduction operations. Annotations have been adjusted to demonstrate use cases and effective strategies when combining these methods. This is an advanced exploration that further expands our knowledge in array manipulation in JavaScript.