Study on Sending Values from DOM to JavaScript in One Bit Code Course
This commit introduces a study on how to send values from the DOM to JavaScript as part of the One Bit Code course. The provided HTML code includes a registration form where users can input a username, password, and confirm the password. Upon clicking the "Register" button, the entered values are sent to the register() function, which validates whether the passwords match and displays a corresponding alert message.