Study on Exercise 12 Resolution in One Bit Code Course
This commit presents a study on the resolution of Exercise 12 in the One Bit Code course. The provided HTML code contains a form for registering developers, including fields for full name and technologies. JavaScript functionality has been implemented to dynamically add input fields for technology name and experience upon clicking the "Add New" button. Upon form submission, the data is collected, and a new developer object is created and added to an array. A success message is displayed, and the form fields are cleared.