Study on Rest Parameters - One Bit Code
This study delves into the usage of rest parameters in JavaScript, as taught in the One Bit Code course. The provided code showcases how rest parameters allow functions to accept an indefinite number of arguments, which are then captured as an array within the function body. Examples demonstrate how rest parameters simplify the handling of varying argument lengths, offering flexibility and convenience in function definitions.