Study on CommonJS Modules - One Bit Code
This commit provides an exploration of CommonJS modules in the One Bit Code course. The code demonstrates the usage of CommonJS modules to organize and modularize JavaScript code. Two modules, render.js and store.js, are imported into the main script using the require function. Each module encapsulates specific functionality related to rendering the application interface and storing information in a database. The export module.exports is utilized to make the functions available for use in other modules.