Solving Exercise 16 - One Bit Code
This commit provides the resolution for Exercise 16. It includes a function called `birthday` that calculates the age of a person based on their birthdate and displays information about their upcoming birthday. The function utilizes the `dayjs` library to work with dates. It calculates the age in years, determines the date of the next birthday, and calculates the number of days until the next birthday. The function is then called with a sample birthdate ('2004-08-20') to demonstrate its functionality.